restructured folder architecture
