+++ package.json
@@ -5,14 +5,15 @@
   "main": "src/server.js",
   "type": "module",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "start": "node src/server.js",
-    "dev": "nodemon src/server.js"
+    "dev": "nodemon src/server.js",
+    "watch": "nodemon -e js,ejs,css"
   },
   "author": "Okoro Ohwo Obed",
   "license": "ISC",
   "dependencies": {
     "bcrypt": "^5.1.0",
+    "connect-flash": "^0.1.1",
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.3",
     "ejs": "^3.1.10",
@@ -24,5 +25,9 @@
     "nodemailer": "^6.9.7",
     "winston": "^3.17.0",
     "winston-timestamp-colorize": "^1.1.0"
+  },
+  "devDependencies": {
+    "connect-livereload": "^0.6.1",
+    "livereload": "^0.9.3"
   }
 }
