+++ src/services/admin/model/admin.js
@@ -1,0 +1,28 @@
+import mongoose from 'mongoose';
+
+const adminSchema = new mongoose.Schema({
+    lastName: {
+        type: String,
+        required: true
+    },
+    firstName: {
+        type: String,
+        required: true
+    },
+    email: {
+        type: String,
+        required: true,
+        unique: true,
+        lowercase: true
+    },
+    password: {
+        type: String,
+        required: true
+    }
+}, {timestamps: true});
+
+// DB INDEXING
+adminSchema.index({ email: 1, password: 1 });
+
+const Admin = mongoose.model('Admin', adminSchema);
+export default Admin;
