+++ src/services/librarian/model/librarian.js
@@ -1,0 +1,35 @@
+import mongoose from 'mongoose';
+import bcrypt from 'bcrypt';
+
+const librarianSchema = new mongoose.Schema({
+    lastName: {
+        type: String,
+        required: true
+    },
+    firstName: {
+        type: String,
+        required: true
+    },
+    email: {
+        type: String,
+        required: true,
+        unique: true,
+        lowercase: true
+    },
+    phone: {
+        type: String,
+        required: true
+    },
+    password: {
+        type: String,
+        required: true
+    }
+}, {timestamps: true});
+
+librarianSchema.pre('save', async function(next) {
+    this.password = await bcrypt.hash(this.password, 10);
+    next();
+});
+
+const Librarian = mongoose.model('Librarian', librarianSchema);
+export default Librarian;
